client/client_requirements.txt
python == 3.6
pycryptodomex == 3.9.7